fix(Onboarding): ajuste de apenas uma onboarding com o usuário após o cadastro e correção de comp duplicados
